requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.7
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
ujson==5.8.0
urllib3==2.0.7
